pyproject.toml
[project]
name = "easydeploy"
version = "0.1.0"
description = "Automation deployment toolkit for robotics AI development environments"
readme = { content-type = "text/plain", text = "Automation deployment toolkit for robotics AI development environments" }
license = { text = "MIT" }
authors = [
    { name = "tyu-mit", email = "tyu@mit.edu" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.0.0",
    "google-cloud-compute>=1.15.0",
    "google-auth>=2.17.0",
    "azure-mgmt-compute>=30.0.0",
    "azure-identity>=1.12.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "paramiko>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
easydeploy = "easydeploy.cli.main:main"
deploy-gcp = "easydeploy.cli.gcp:main"
deploy-azure = "easydeploy.cli.azure:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/easydeploy"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# Enable only essential rules for productivity
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming (relaxed)
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "N806",   # variable in function should be lowercase (too strict)
    "N803",   # argument name should be lowercase (too strict)
]

[tool.ruff.lint.isort]
known-first-party = ["easydeploy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "cloud: Tests requiring cloud credentials",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=6.3.0",
    "ruff>=0.12.12",
]
